+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 Bug fixes:
 
+- Support a named datetime or object dtype index in `explore()` (#3360, #3364).
 - Fix a regression preventing a Series as an argument for geometric methods (#3363)
-- Support a named datetime index in `explore()` (#3360).
 
 ## Version 1.0.0 (June 24, 2024)